Added login framework. Need to add database component
