+++ login_server/server.js
@@ -1,0 +1,5 @@
+const app = require('./src/app.js')
+
+app.listen(3000, () => {
+    console.log('Server is running on port 3000');
+});
